问题详情接口开发

+++ mercury/common/category.go
@@ -1,6 +1,6 @@
 package common
 
 type Category struct {
-	CategoryId   int    `json:"id" db:"category_id"`
+	CategoryId   int64  `json:"id" db:"category_id"`
 	CategoryName string `json:"name" db:"category_name"`
 }
